Version 1.0.2
* Fixed an issue where no plugin hooks would be deployed (causing
nothing generated by the plugin to be visible on the forum) (GitHub
Issue #1).
* Fixed (hopefully) an issue where PHP would throw an "argument is not
an array" error on the profile and post bit pages (GitHub Issue #6).
* Set a timeout on the cURL function used for accessing the Steam
network. This should mean that it should no longer cause your forum to
stop loading (or to take a VERY long time to load) if it is having
difficulty contacting Steam.